Refactor setup_directories_and_models.py, move to download models and process models, add change to readme and config.ini
