+++ config.ini
@@ -19,7 +19,7 @@ clip_model_dir = ${MODELS_DIRS:CLIP_MODELS_DIR}${BASE:CLIP_MODEL_NAME}/
 text_embedder_dir = ${MODELS_DIRS:CLIP_MODELS_DIR}text_embedder/
 image_encoder_dir = ${MODELS_DIRS:CLIP_MODELS_DIR}image_encoder/
 
-[SUBMODELS_DIRS]
+[CLIP_MODELS_DIRS]
 tokenizer_dir = ${MODELS_DIRS:TEXT_EMBEDDER_DIR}tokenizer/
 text_model_dir = ${MODELS_DIRS:TEXT_EMBEDDER_DIR}text_model/
 image_processor_dir = ${MODELS_DIRS:IMAGE_ENCODER_DIR}image_processor/
@@ -31,11 +31,4 @@ unet_path = ${MODELS_DIRS:SD_DEFAULT_MODEL_DIR}unet.safetensors
 autoencoder_path = ${MODELS_DIRS:SD_DEFAULT_MODEL_DIR}autoencoder.safetensors
 latent_diffusion_path = ${MODELS_DIRS:SD_DEFAULT_MODEL_DIR}latent_diffusion.safetensors
 
-[CLIP_PATHS]
-image_processor_path = ${MODELS_DIRS:IMAGE_ENCODER_DIR}image_processor
-vision_model_path = ${MODELS_DIRS:IMAGE_ENCODER_DIR}vision_model
-image_encoder_path = ${MODELS_DIRS:IMAGE_ENCODER_DIR}image_encoder.safetensors
-tokenizer_path = ${MODELS_DIRS:TEXT_EMBEDDER_DIR}tokenizer
-text_model_path = ${MODELS_DIRS:TEXT_EMBEDDER_DIR}text_model
-text_embedder_path = ${MODELS_DIRS:TEXT_EMBEDDER_DIR}text_embedder.safetensors
